% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{led_bulb}
\alias{led_bulb}
\title{Select a random sample of bulbs all time points.}
\usage{
led_bulb(
  number_bulbs = 1,
  seed = 9012007,
  dat = data4led::led_study,
  columns_keep = c("id", "hours", "intensity", "percent_intensity")
)
}
\arguments{
\item{number_bulbs}{The number of bulbs to sample. Defaults to 1.}

\item{seed}{is used to make the random sample reproducible.  May change at teacher request.}

\item{dat}{is the data object from which the sample is taken.  Defaults to led_study.}

\item{columns_keep}{are the columns that should be returned from dat.}
}
\description{
A function to standardize the random selection of bulbs for the LED project in BYUI's MATH 119.
}
\examples{
led_bulb(2, seed = 1976)
}
