% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{led_histogram}
\alias{led_histogram}
\title{Create a histogram for bulb values.}
\usage{
led_histogram(led_data, use_intensity = FALSE, nbins = 15)
}
\arguments{
\item{led_data}{is a data object from the function led_bulb().}

\item{use_intensity}{defines whether inensity or percent intensity is plotted on the y-axis.}

\item{nbins}{the number of bins to use in the histogram}
}
\description{
Plot a ggplot2 histogram.
}
\examples{
led_histogram(led_time(425))
}
