% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{led_linechart}
\alias{led_linechart}
\title{Create a time-series chart of LED bulbs.}
\usage{
led_linechart(led_data, use_intensity = FALSE)
}
\arguments{
\item{led_data}{is a data object from the function led_bulb().}

\item{use_intensity}{defines whether inensity or percent intensity is plotted on the y-axis.}
}
\description{
Plot a ggplot2 linechart colored by bulb.
}
\examples{
led_linechart(led_bulb(5))
led_linechart(led_bulb(5), TRUE)
}
