% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{led_testing}
\alias{led_testing}
\title{LED example bulbs of lumen output for two products with standard procedure time point measurements}
\format{
A data frame with columns:
\describe{
 \item{id}{The variable is numeric. An id for each LED light measured}
 \item{hours}{The variable is numeric. The number of hours since the first measurement}
 \item{percent_intensity}{The variable is numeric. The percent light output based on the first measured intensity of the bulb}
 \item{company}{The variable is character. Either A or B to represent to different company products}
}
}
\source{
\url{data_led_private}
}
\usage{
led_testing
}
\description{
An example data set of LED bulbs based on actual data.
}
\examples{
\dontrun{
led_testing
}
}
\keyword{datasets}
