% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{led_time}
\alias{led_time}
\title{Select all bulbs at the nearest time.}
\usage{
led_time(
  time_bulb = 294,
  number_bulbs = NULL,
  seed = 9012017,
  dat = data4led::led_study,
  columns_keep = c("id", "hours", "intensity", "percent_intensity")
)
}
\arguments{
\item{time_bulb}{The bulb time that you would like. The most changed argument to the function.}

\item{number_bulbs}{The number of bulbs to sample. Defaults to all.}

\item{seed}{is used to make the random sample reproducible.  May change at teacher request.}

\item{dat}{is the data object from which the sample is taken.  Defaults to led_study.}

\item{columns_keep}{are the columns that should be returned from dat.}
}
\description{
A function to standardize the time point selection for all LED bulbs for the LED project in BYUI's MATH 119.
}
\examples{
led_time(425, number_bulbs = 10, seed = 1976)
}
